Add new Generator and Generatable types. Implements #152
This commit adds two new types to implement something like Iterable/Iterator for generated, infinite series. It also adds a Matcher which tests the start sequence of Generabtables.
In addition it adds initial implementations for Sequences which are based on an initial value and a Function which calculates the next value based on that.

In order to use Generatables like Iterables (which needs to be done carefully because they are infinite) there are initial Generatable-Iterable and Generator-Iterator adapters.