Introduce simplified Optional implements #149 (#158)
This introduces a new, simplified `Optional` type in the jems package which doesn't have the `value(T default)` method. This method caused a lot of duplicated code and can be replaced with a `Backed` single.
In addition it duplicates all (now deprecated) `Optional` implementations in the jems package.

Note, for compatibility, some of the implementations in the jems package still use the old `Optional`. This is subject to change.